+++ integration-tests/tests/src/hooks/open-realm-before.ts
@@ -16,7 +16,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-import Realm from "realm";
+import Realm, { User } from "realm";
 
 import { openRealm, OpenRealmConfiguration } from "../utils/open-realm";
 
@@ -29,9 +29,7 @@ import { openRealm, OpenRealmConfiguration } from "../utils/open-realm";
  * @returns Promise which resolves when complete
  */
 export function openRealmHook(config: OpenRealmConfiguration = {}) {
-  return async function openRealmHandler(
-    this: Partial<RealmContext> & Partial<UserContext> & Mocha.Context,
-  ): Promise<void> {
+  return async function openRealmHandler(this: Partial<RealmContext> & UserContext & Mocha.Context): Promise<void> {
     this.longTimeout();
     if (this.realm) {
       throw new Error("Unexpected realm on context, use only one openRealmBefore per test");
@@ -39,7 +37,7 @@ export function openRealmHook(config: OpenRealmConfiguration = {}) {
       this.closeRealm = async () => {
         console.warn("🤷 Skipped closing a Realm that failed to open");
       };
-      const { realm, config: actualConfig } = await openRealm(config, this.user);
+      const { realm, config: actualConfig } = await openRealm(config, this.user as unknown as User);
       this.realm = realm;
       this.closeRealm = async ({
         clearTestState = true,
@@ -58,7 +56,7 @@ export function openRealmHook(config: OpenRealmConfiguration = {}) {
           Realm.clearTestState();
         }
         if (reopen) {
-          const { realm } = await openRealm(actualConfig, this.user);
+          const { realm } = await openRealm(actualConfig, this.user as unknown as User);
           this.realm = realm;
         }
       };