+++ integration-tests/tests/src/tests/dictionary.ts
@@ -52,7 +52,7 @@ const EmbeddedChild = {
   },
 };
 
-const DictTypedSchema = {
+const DictTypedSchema: Realm.ObjectSchema = {
   name: "TypedDictionary",
   properties: {
     dict1: { type: "dictionary", objectType: "Children" }, // dictionary of objects is nullable by default
@@ -363,15 +363,16 @@ describe("Dictionary", () => {
       expect(stringifiedAndParsed).deep.equals({ dict: values });
     });
 
-    // TODO: Unskip once https://github.com/realm/realm-core/issues/4805 is fixed
-    it.skip("throws a meaningful error if accessed after deletion", function (this: RealmContext) {
+    it("throws a meaningful error if accessed after deletion", function (this: RealmContext) {
       this.realm.write(() => {
         const item = this.realm.create<Item>("Item", {});
         const dict = item.dict;
         this.realm.delete(item);
         expect(() => {
           JSON.stringify(dict);
-        }).throws("Access to invalidated Dictionary object");
+        }).throws(
+          "Dictionary is no longer valid. Either the parent object was deleted or the containing Realm has been invalidated or closed.",
+        );
       });
     });