+++ integration-tests/tests/src/tests/listeners.ts
@@ -66,6 +66,7 @@ describe("Realm Listeners", () => {
       });
 
       this.realm.write(() => {
+        //@ts-expect-error Internal method
         this.realm._updateSchema([...this.realm.schema, { name: "MyClass", properties: { myField: "string" } }]);
       });
     });
@@ -126,6 +127,7 @@ describe("Realm Listeners", () => {
         }
       });
       this.realm.write(() => {
+        //@ts-expect-error Internal method
         this.realm._updateSchema([...this.realm.schema, { name: "MyClass", properties: { myField: "string" } }]);
       });
       this.realm.write(() => {
@@ -168,6 +170,7 @@ describe("Realm Listeners", () => {
         this.realm.create("Person", { age: 3, name: "Tom" });
       });
       this.realm.write(() => {
+        //@ts-expect-error Internal method
         this.realm._updateSchema([...this.realm.schema, { name: "MyClass", properties: { myField: "string" } }]);
       });
     });