+++ integration-tests/tests/src/tests/sync/client-reset.ts
@@ -17,7 +17,16 @@
 ////////////////////////////////////////////////////////////////////////////
 
 import { expect } from "chai";
-import Realm, { ClientResetMode, SessionStopPolicy, BSON } from "realm";
+import Realm, {
+  BSON,
+  ClientResetMode,
+  SessionStopPolicy,
+  User,
+  App,
+  Configuration,
+  ConfigurationWithSync,
+  SyncError,
+} from "realm";
 import { authenticateUserBefore, importAppBefore } from "../../hooks";
 import { DogSchema, PersonSchema } from "../../schemas/person-and-dog-with-object-ids";
 import { expectClientResetError } from "../../utils/expect-sync-error";
@@ -34,7 +43,7 @@ const FlexibleDogSchema = { ...DogSchema, properties: { ...DogSchema.properties,
  */
 function addSubscriptions(realm: Realm): void {
   const subs = realm.subscriptions;
-  subs.update((mutableSubs: Realm.App.Sync.MutableSubscriptionSet) => {
+  subs.update((mutableSubs) => {
     mutableSubs.add(realm.objects(FlexiblePersonSchema.name));
     mutableSubs.add(realm.objects(FlexibleDogSchema.name));
   });
@@ -59,7 +68,7 @@ async function waitServerSideClientResetDiscardUnsyncedChangesCallbacks(
   const resetHandle = createPromiseHandle();
   let afterCalled = false;
   let beforeCalled = false;
-  const config: Configuration = {
+  const config: ConfigurationWithSync = {
     schema,
     sync: {
       user,
@@ -107,7 +116,7 @@ async function waitServerSideClientResetRecoveryCallbacks(
   let afterCalled = false;
   let beforeCalled = false;
 
-  const config: Configuration = {
+  const config: ConfigurationWithSync = {
     schema,
     sync: {
       user,
@@ -154,7 +163,7 @@ async function waitSimulatedClientResetDiscardUnsyncedChangesCallbacks(
   let afterCalled = false;
   let beforeCalled = false;
 
-  const config: Configuration = {
+  const config: ConfigurationWithSync = {
     schema,
     sync: {
       user,
@@ -204,7 +213,7 @@ async function waitSimulatedClientResetRecoverCallbacks(
   let afterCalled = false;
   let beforeCalled = false;
 
-  const config: Configuration = {
+  const config: ConfigurationWithSync = {
     schema,
     sync: {
       user,
@@ -290,7 +299,7 @@ function getSchema(useFlexibleSync: boolean) {
       it(`manual client reset requires either error handler, client reset callback or both (${getPartialTestTitle(
         useFlexibleSync,
       )} sync)`, async function (this: RealmContext) {
-        const config: Configuration = {
+        const config: ConfigurationWithSync = {
           schema: getSchema(useFlexibleSync),
           sync: {
             // @ts-expect-error this setting is not for users to consume
@@ -309,18 +318,20 @@ function getSchema(useFlexibleSync: boolean) {
       it(`handles manual simulated client resets with ${getPartialTestTitle(
         useFlexibleSync,
       )} sync enabled`, async function (this: RealmContext) {
-        await expectClientResetError(
-          {
-            schema: getSchema(useFlexibleSync),
-            sync: {
-              _sessionStopPolicy: SessionStopPolicy.Immediately,
-              ...(useFlexibleSync ? { flexible: true } : { partitionValue: getPartitionValue() }),
-              user: this.user,
-              clientReset: {
-                mode: ClientResetMode.Manual,
-              },
+        const config: ConfigurationWithSync = {
+          schema: getSchema(useFlexibleSync),
+          sync: {
+            //@ts-expect-error Internal field
+            _sessionStopPolicy: SessionStopPolicy.Immediately,
+            ...(useFlexibleSync ? { flexible: true } : { partitionValue: getPartitionValue() }),
+            user: this.user,
+            clientReset: {
+              mode: ClientResetMode.Manual,
             },
           },
+        };
+        await expectClientResetError(
+          config,
           this.user,
           (realm) => {
             if (useFlexibleSync) {
@@ -331,7 +342,7 @@ function getSchema(useFlexibleSync: boolean) {
             // @ts-ignore calling undocumented method _simulateError
             session._simulateError(211, "Simulate Client Reset", "realm::sync::ProtocolError", false); // 211 -> diverging histories
           },
-          (error) => {
+          (error: SyncError) => {
             expect(error.name).to.equal("ClientReset");
             expect(error.message).to.equal("Simulate Client Reset");
             expect(error.code).to.equal(211);
@@ -342,8 +353,8 @@ function getSchema(useFlexibleSync: boolean) {
       it(`handles manual simulated client resets by callback with ${getPartialTestTitle(
         useFlexibleSync,
       )} sync enabled`, async function (this: RealmContext) {
-        return new Promise<void>((resolve, _) => {
-          const config: Configuration = {
+        return new Promise<void>((resolve) => {
+          const config: ConfigurationWithSync = {
             schema: getSchema(useFlexibleSync),
             sync: {
               // @ts-expect-error this setting is not for users to consume
@@ -352,7 +363,7 @@ function getSchema(useFlexibleSync: boolean) {
               user: this.user,
               clientReset: {
                 mode: ClientResetMode.Manual,
-                onManual: (session: Realm.App.Sync.Session, path: string) => {
+                onManual: (session, path) => {
                   expect(session).to.be.not.null;
                   expect(path).to.not.empty;
                   resolve();
@@ -376,7 +387,7 @@ function getSchema(useFlexibleSync: boolean) {
         useFlexibleSync,
       )} sync enabled`, async function (this: RealmContext) {
         return new Promise((resolve, reject) => {
-          const config: Configuration = {
+          const config: ConfigurationWithSync = {
             schema: getSchema(useFlexibleSync),
             sync: {
               // @ts-expect-error this setting is not for users to consume