+++ integration-tests/tests/src/tests/sync/open-behavior.ts
@@ -105,17 +105,16 @@ describe("OpenBehaviour", function () {
     });
     realm.close();
 
-    const config2: Realm.Configuration = {
+    const config2 = {
       schema: [DogForSyncSchema],
       sync: {
-        user,
+        user: user as Realm.User,
         partitionValue,
-        //@ts-expect-error const enum is removed at runtime, should either remove const or leave this as is.
         existingRealmFileBehavior: { type: "openImmediately" },
       },
     };
 
-    realm = await Realm.open(config2);
+    realm = await Realm.open(config2 as Realm.Configuration);
 
     expect(realm.objects(DogForSyncSchema.name).length).equals(1);
   });
@@ -129,9 +128,8 @@ describe("OpenBehaviour", function () {
       sync: {
         user,
         partitionValue,
-        //@ts-expect-error const enum is removed at runtime, should either remove const or leave this as is.
+        //@ts-expect-error TYPEBUG: cannot access const enum at runtime
         _sessionStopPolicy: "immediately",
-        //@ts-expect-error const enum is removed at runtime, should either remove const or leave this as is.
         newRealmFileBehavior: { type: "downloadBeforeOpen" },
       },
     };
@@ -190,7 +188,6 @@ describe("OpenBehaviour", function () {
           partitionValue,
           //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           _sessionStopPolicy: "immediately",
-          //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           existingRealmFileBehavior: { type: "downloadBeforeOpen" },
         },
       };
@@ -212,10 +209,8 @@ describe("OpenBehaviour", function () {
         //@ts-expect-error TYPEBUG: cannot access const enum at runtime
         _sessionStopPolicy: "immediately",
         newRealmFileBehavior: {
-          //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           type: "downloadBeforeOpen",
           timeOut: 0,
-          //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           timeOutBehavior: "throwException",
         },
       },
@@ -258,10 +253,8 @@ describe("OpenBehaviour", function () {
           //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           _sessionStopPolicy: "immediately",
           existingRealmFileBehavior: {
-            //@ts-expect-error TYPEBUG: cannot access const enum at runtime
             type: "downloadBeforeOpen",
             timeOut: 0,
-            //@ts-expect-error TYPEBUG: cannot access const enum at runtime
             timeOutBehavior: "throwException",
           },
         },
@@ -321,10 +314,8 @@ describe("OpenBehaviour", function () {
           //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           _sessionStopPolicy: "immediately",
           newRealmFileBehavior: {
-            //@ts-expect-error TYPEBUG: cannot access const enum at runtime
             type: "downloadBeforeOpen",
             timeOut: 0,
-            //@ts-expect-error TYPEBUG: cannot access const enum at runtime
             timeOutBehavior: "openLocalRealm",
           },
         },
@@ -402,10 +393,8 @@ describe("OpenBehaviour", function () {
           //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           _sessionStopPolicy: "immediately",
           existingRealmFileBehavior: {
-            //@ts-expect-error TYPEBUG: cannot access const enum at runtime
             type: "downloadBeforeOpen",
             timeOut: 0,
-            //@ts-expect-error TYPEBUG: cannot access const enum at runtime
             timeOutBehavior: "openLocalRealm",
           },
         },
@@ -466,14 +455,10 @@ describe("OpenBehaviour", function () {
         sync: {
           user,
           partitionValue,
-          //@ts-expect-error TYPEBUG: cannot access const enum at runtime
-          _sessionStopPolicy: "immediately",
           existingRealmFileBehavior: {
-            //@ts-expect-error TYPEBUG: cannot access const enum at runtime
-            type: "downloadBeforeOpen",
+            type: OpenRealmBehaviorType.DownloadBeforeOpen,
             timeOut: 1000,
-            //@ts-expect-error TYPEBUG: cannot access const enum at runtime
-            timeOutBehavior: "openLocalRealm",
+            timeOutBehavior: OpenRealmTimeOutBehavior.OpenLocalRealm,
           },
         },
       };
@@ -497,7 +482,6 @@ describe("OpenBehaviour", function () {
         partitionValue,
         //@ts-expect-error TYPEBUG: cannot access const enum at runtime
         _sessionStopPolicy: "immediately",
-        //@ts-expect-error TYPEBUG: cannot access const enum at runtime
         newRealmFileBehavior: { type: "downloadBeforeOpen" },
       },
     };
@@ -532,7 +516,6 @@ describe("OpenBehaviour", function () {
         partitionValue,
         //@ts-expect-error TYPEBUG: cannot access const enum at runtime
         _sessionStopPolicy: "immediately",
-        //@ts-expect-error TYPEBUG: cannot access const enum at runtime
         newRealmFileBehavior: { type: "downloadBeforeOpen" },
       },
     };
@@ -555,7 +538,6 @@ describe("OpenBehaviour", function () {
         partitionValue,
         //@ts-expect-error TYPEBUG: cannot access const enum at runtime
         _sessionStopPolicy: "immediately",
-        //@ts-expect-error TYPEBUG: cannot access const enum at runtime
         newRealmFileBehavior: { type: "downloadBeforeOpen" },
       },
     };
@@ -595,12 +577,10 @@ describe("OpenBehaviour", function () {
         sync: {
           user,
           partitionValue,
-          //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           _sessionStopPolicy: "immediately",
-          //@ts-expect-error testing invalid value for newRealmFileBehaviour
           newRealmFileBehavior: { type: "foo" },
         },
-      }),
+      } as unknown as Realm.Configuration),
     ).throws;
 
     await expect(() =>
@@ -609,12 +589,10 @@ describe("OpenBehaviour", function () {
         sync: {
           user,
           partitionValue,
-          //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           _sessionStopPolicy: "immediately",
-          //@ts-expect-error testing invalid value for newRealmFileBehaviour
-          newRealmFileBehavior: { type: "openLocalRealm", timeOutBehavior: "foo" },
+          newRealmFileBehavior: { type: OpenRealmBehaviorType.OpenImmediately, timeOutBehavior: "foo" },
         },
-      }),
+      } as unknown as Realm.Configuration),
     ).throws;
 
     await expect(() =>
@@ -623,12 +601,10 @@ describe("OpenBehaviour", function () {
         sync: {
           user,
           partitionValue,
-          //@ts-expect-error TYPEBUG: cannot access const enum at runtime
           _sessionStopPolicy: "immediately",
-          //@ts-expect-error testing invalid value for newRealmFileBehaviour
-          newRealmFileBehavior: { type: "openLocalRealm", timeOut: "bar" },
+          newRealmFileBehavior: { type: OpenRealmBehaviorType.OpenImmediately, timeOut: "bar" },
         },
-      }),
+      } as unknown as Realm.Configuration),
     ).throws;
   });
 });